requirements.txt
langchain




streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-huggingface
validators